swagger clean up config
