+++ greeting-service/src/main/resources/application.properties
@@ -1,2 +1,3 @@
 spring.application.name=say-hello
 server.port=8080
+
